modify GetScoreRk algorithm to be linear in # of fields
Previous implementation had exponential potential combinations w.r.t.
the number of fields in RankKey. This new implementation instead checks
each field individually per map key for performance cost that is linear
to the number of fields. (This only works because the number of mappings
per import name can be assumed to be constant).