change RankKey fields to be *string (from string)
We want to remove the placeholder "*" while still allowing for empty
strings to be valid matchers in StatsRanking. We are not able to keep
RankKey as a string type and satisfy that requirement at the same time
because the zero value of the string type in Go is the empty string
`""`. Instead, we make the RankKey fields be pointers and update the
code accordingly to work with pointers.

- Shortened names of RankKey fields to make it easier for StatsRanking
- Modified GetScoreRk to check for pointers nilness and dereference
  non-nill fields when comparing
- Modified the GetScore* adapters to pass string references
- Updated StatsRanking to use the new style and not have `*`
  placeholders.
- Added a function s(str string) *string that returns a pointer to the
  string passed to it and used it as a helper in StatsRanking.
- Updated comments according to the new changes