Redesign ranking scoring algorithm (#866)
* add GetScoreRk which uses the existing scoring algorithm

The current algorithm is duplicated three times, I will
use this function to DRY that. RankKey is used as a generic
holder of RankKey related information.

* reduce GetScore* duplication using GetScoreRk

* update comments in ranking.go

indicate that GetScoreRk is the main scoring algorithm, and refer
to it from the "adapter" methods.

* extract ImportName from RankKey

Previously, ImportName was a field of RankKey like measurementMethod,
etc. This change removes ImportName from RankKey, and instead makes it
an outer key for StatsRanking. This includes the following changes:

- StatsRanking signature change from "map[RankKey]int" to
  "map[string]map[RankKey]int"
- Update the definitions of StatsRanking such that the import names
  are extracted and become the outer key
    - In the two cases where we have >1 "RankKey:score" mappings per
      importName, merge them under the same outer key.
- Update GetScoreRk() signature to take "importName string" as the first
  input
    - Update calls to GetScoreRk from the GetScore* adapters such that
      the import name is passed as a parameter to GetScoreRk and not as
      a field of RankKey
- In the GetScoreRank() algorithm, "fail fast" by returning BaseRank if
  import name does not exist in StatsRanking

* modify GetScoreRk algorithm to be linear in # of fields

Previous implementation had exponential potential combinations w.r.t.
the number of fields in RankKey. This new implementation instead checks
each field individually per map key for performance cost that is linear
to the number of fields. (This only works because the number of mappings
per import name can be assumed to be constant).

* clarify GetScoreRk if clause with helper variables

* codacy changes

* nit: typo

* nit: s/scoreOfMostExactMatch/rankScore

* break up complex condition to linear, concise logic

* change RankKey fields to be *string (from string)

We want to remove the placeholder "*" while still allowing for empty
strings to be valid matchers in StatsRanking. We are not able to keep
RankKey as a string type and satisfy that requirement at the same time
because the zero value of the string type in Go is the empty string
`""`. Instead, we make the RankKey fields be pointers and update the
code accordingly to work with pointers.

- Shortened names of RankKey fields to make it easier for StatsRanking
- Modified GetScoreRk to check for pointers nilness and dereference
  non-nill fields when comparing
- Modified the GetScore* adapters to pass string references
- Updated StatsRanking to use the new style and not have `*`
  placeholders.
- Added a function s(str string) *string that returns a pointer to the
  string passed to it and used it as a helper in StatsRanking.
- Updated comments according to the new changes

* add comments and use s() instead of & in GetScore*