Add a loadtesting tool based on BulkObservationsSeriesLinked rpc (#889)
Adding the tool I used to loadtest some of the APIs. I think it will be useful for future testing. 

The easiest way to loadtest apis that are not exposed to the internet was to deploy the locust pod and specify the k8s service name as the host.

Below is a screenshot of the locust UI. 

<img width="1418" alt="Screen Shot 2022-08-23 at 8 39 14 AM" src="https://user-images.githubusercontent.com/8507196/186201632-e5341454-d380-46da-98e6-9fb059aba71e.png">

Co-authored-by: Alex Chen <alexyfchen@google.com>