Update API naming convention for entity/node/dcid (#900)
"node" represents a knowledge graph node which could be of any kind
while "entity" should only be used for real world entities. This PR
updates the API parameters and code usage to follow this pattern:

* "node" is used for node APIs: properties, property values and triples
* "entity" is used for observation APIs and it's mostly used together
with "variable". This is current behavior and not changed in this PR.
* "dcid" is used for specialized API like /info/* where the actual
entity type is already clear in the API name, so using "dcid" is better
than "entity"

All these changes only affect website clients, which will be updated in
one PR with mixer submodule update.