Implement DerivedObservationsSeries for on-the-fly SVCalculator. (#910)
This is a first version of the on-the-fly SVCalculator.

Lots of logic are focused around parsing the formula (with optional
filters) with Golang's AST package, as well as recursively iterate
through AST to extract information and perform series computation.

The main caveat/TODO is: No fancy ranking for facet yet -- After
necessary filtering, if there are still multiple series/facets, the code
just concatenate the facet contents and compare the string, and pick the
largest one deterministically. That being said, I think this PR is large
enough for a first version.

The design doc will also be updated to better illustrate the algorithms.