Support dynamically created Bigtable client for data read (#923)
This allows users to specify additional custom built import groups in
API request.

A bigtable client is created on the fly in the request context. The user
provided import group ranks the highest in all the import groups.

A test cache is built in "datcom-mixer-website" project and used in
golden tests. This PR also updated observation series API to use the
custom import group params as an example.