Add V1 API for getting observation dates based on entity count (#940)
Compared with previous version, the major change here is the response
shape, where it's oriented by the date with the entity frequency for
each facet. This aligns well with its usage, where user wants to know
which facet to use for each date.

The dates and count for facets are both sorted.

Also updated the uri and args to match V1 patterns.