Treat recon APIs as regular mixer API and remove Recon service (#962)
Recently Recon server is merged in to Mixer server and the recon API
requests are routed into dedicated GKE pods. With these changes, it is
not necessary to have a separate Recon service anymore.

This cleanup simplifies overall deployment logic and make it easier to
extend to various custom DC cases.