Remove two deprecated API: GetStatSetSeries, GetStatSetSeriesWithinPlace (#971)
These two APIs are never made public and were only used for the website
app. As the website migrated to V1 APIs, these can be removed.

To confirm on the usage. Here are the traffic:

Mixer (30 days):

![image](https://user-images.githubusercontent.com/5951856/200053219-75c345d8-1763-402c-a86f-55b69d7ef826.png)

Website (7 days, as the usage was removed since last release):

![image](https://user-images.githubusercontent.com/5951856/200053472-730050d0-09c9-43ee-a54a-746419819686.png)

The 4 calls to /stat/set/series/within-place have IP from Google
crawlers.