Clean up docs about minikube, skaffold and kustomize (#1083)
Bringing up local Kubernetes cluster with Minikube is slow, subject to
memory limit and is very slow. Using envoy is a much better choice for
testing API locally. So no need to keep minikube (with skaffold) as an
option for local development.

Also remove reference to Kustomize as it's replaced by Helm in mixer
already.