[V2 API] A unified observation API for fetching stat values and non-stat properties (#1086)
In V2, I try to have a unified API to query observation stat value and
metrics. This PR is an attempt to upgrade V1 GetPlaceVariables and
CheckObservationExistence API.

**GetPlaceVariables**
In V1, the response is a map keyed by stat var with values being a list
of entity dcids.

**CheckExistence**
In V1, the response is a two level map with keys being variable and
entity, and value being a boolean.

In V2, they are all converted into GetObservation() and make the API
argument with SQL style.

1. The request has a new "select" fields to specify what to return. This
makes it possible to use one API to fetch stat values and other
properties like date and variables.

2. To get place variables, need to loop through the keys of the
response, instead of a plain list. The complexity for client is similar.
When multiple entities are provided, there is a "union" bit in V1 which
is inherently supported (and with finer granularity) in V2: Each
variable contains all the places that has observations (see the golden
response for detail)

This also updated the response field from "observations_by_variable" to
"by_variable" so it also suits to the Observation Metric API. It's
concise and still meaning full with "observations" as the inner fields
in GetObservation() API response.