Optimize when to read time series cache for contained in observation query (#1139)
Copies the BT build rule directly to determine if collection cache
exists for a given query. Now the time series cache is read only when
needed.

This saves a lot of processing for places with no data. In some cases,
when a lot of variables and child places are involved, OOM can happen
but there is no need to fetch the time series data at all.