Fix mixer bigtable read context cancel issue (#1220)
This issue is uncovered from observation series BT cache read where
O(100) variable and O(10) places are queried. In this case, the context
used to read Bigtable is canceled after certain amount of read call.
This won't be an issue when the variable and place number is small.

In the previous implementation, each Bigtable table has its own client
to the instance, thus there are 10+ client from one binary connected to
the instance. With this change, there is only one shared client for all
the base bigtable (one for each import group). This apparently solves
the issue.