RecognizePlace: Implement adjectival place qualification (#1249)
We want to interpret China as a country in queries like [population
across chinese provinces], but not in queries like [chinese speakers in
california].

To do this, we qualify these "adjectival" usages only if followed
immediately by place types. In this case, we leave the place type in the
original query so NL can consume it.

The adjectival words were largely from
https://en.wikipedia.org/wiki/List_of_adjectival_and_demonymic_forms_for_countries_and_nations.
On that I ran place-name-resolver to map to DCID and skimmed thru the
mappings.