[recognize entities] handle common words (#1349)
There are some reconName values that are very common words like "me",
"to", "a". To handle these, added a map of reconNam value to the allowed
types of the entities recognized & only recognize a name if the name is
preceded by its type in the query.

For example, there is a gene called "Me". The query "Tell me about the
gene me" should not recognize the first "me", but should recognize the
"gene me".