Add basic support for calculations (#1364)
This also splits up the fetching and merging for the formulas (follow up
to https://github.com/datacommonsorg/mixer/pull/1362) and adds basic
tests.

This currently does not support matching facets, which will be a follow
up.

Currently, this independently calculates each missing variable/entity
pair, rather than combining the entities into one DerivedSeries request,
and then updates the merging to accept an array of responses. This was
to avoid computing the calculation if some of the entities already have
data, however if it's preferred to only do one DerivedSeries per formula
I can update this.