Add pagination to formula_fetcher (#1424)
Fixes a bug where the formula_fetcher previously did not include
pagination, so only retrieved the first page of formulas. We recently
added several hundred more formulas, so need to use pagination to
retrieve them all. As part of this change, switches the formula_fetcher
to use the triples package directly to make pagination simpler.

This fixes some goldens that had been empty due to the bug before.