Set is_dc_aggregate for calculated responses (#1421)
This is so the output facet will be different than the input facets and
the response indicates that the result is aggregated.

This will also modify the facet ids for custom import calculations

UPDATE: Based on offline feedback, this also modifies the facet id
construction to only include fields that have been set. This is to help
keep the ids stable even as fields are potentially added in the future.