Allow Combining a 5-token span to account for the formatting of the Google Maps predictions API. (#1434)
Today, a 3 part place does not work appropriately, the 3rd containing
part is considered its own location. Example
https://screenshot.googleplex.com/3qcq4dB3zKp6Qte

The Google Maps prediction API being added for autocomplete uses a
3-part format for cities in countries that use states, US, Brazil,
Argentina and more.

With this change, we can support a query autocompleted such as:
https://screenshot.googleplex.com/4jnCABAFVTHNTUs

We intentionally did NOT account for "San Diego, California USA". The
Google Maps API includes the commas so figured it would not be worth the
code clutter to account for a comma ommission.

Tested locally + added a unit test.