Updates to V3 observation (#1503)
This includes: 
- Ranking facets in the "preferred" order
- Updates to the response structure. This is primarily to match V2 to
help with merging and backward compatibility. V2 is a bit inconsistent
with the responses for different request patterns, particularly for
missing data and containedInPlace (likely due to using different BT
caches for each), so for this PR I similarly handled the observation,
facet, and existence response constructions differently.

Going forward, I think it would be nice to unify these responses more,
which will simplify the V3 implementation (and is possible with
Spanner!) but I did see some website usage that looks like it could
depend on the current structure, so didn't want to change anything now.

As a follow up, I'll add datasource merging and series limits, and then
(I think?) observation should be complete. (I'll test this more though)