Add id field to pagination proto (#1531)
Originally was trying to reuse the V1 pagination proto, but it was very
specific to BT/import group implementation so was difficult to use for
Spanner. So this will hopefully be more generic, and each Data Source
can have it's own implementation.

As a side effect, I needed to modify the remote datasource to use V3
endpoint, since V2 relies on v1Pagination.remotePaginationInfo (but now
we won't be distinguishing remote vs other data source)

Edit: now this just adds a new field for the spanner implementation