Add entity existence requests to V3 observation (#1557)
This is for requests of the form /observation?entity.dcids=<> where no
variable is specified, so essentially getting the variables for a
place(s).

Kara brought up that this is supported in V2, but we noticed it wasn't
yet supported it in V3. In V2, this only works for existence and not
with any date or facet filters (so it's a bit inconsistent). Here I
added it everywhere, but can restrict it if needed as a follow up, along
with some other query optimizations

This PR also splits up some tests into different tests to avoid test
timeouts