Upgrade golangci-lint to 2.3.0 and fix or ignore existing issues (#1581)
The upcoming upgrade of golang version causes new existing errors to
surface.

While here, I updated `run_test.sh` to be similar to the version in the
website repo: located in the repo root and capable of running or
auto-fixing with the linter.

There is also one resolve test that fails for me even at head. Not sure
how that made it through previous precommit hooks, but I updated the
golden.