Instrument cache reads with OpenTelemetry metrics (#1583)
- Add a `metrics_exporter` flag to Mixer server that switches between
metrics export modes, with blank = none
- Add a new metrics package with documentation for local debugging
- Add the Google-built OTLP collector to the deploy process using a
sub-script that should be reusable for the website's mixer with
subsequent changes to the website deploy script
- With `--metrics_exporter=otlp` and this collector job deployed, the
mixer server will regularly push metrics via gRPC to the collector, and
from there they are exported to Google-managed Prometheus which surfaces
them in Cloud Monitoring.
- Enable metrics in dev
- Add metric for Bigtable cache read latency
  - Each multi-table fanout counts as a single read
  - The metric can also be used as a count of reads
- Attributes allow grouping/filtering by: RPC method context, cache
prefix, and read outcome
- Add metric for cachedata accesses
- The cachedata struct is loaded from Bigtable, so understanding its
usage is important for eventual migration off BT

Tested by deploying to mixer dev, sending a batch of requests to the dev
API, and examining in Cloud Monitoring.