Implement feature flags with a ConfigMap and Reloader (#1624)
Add a `featureflags` module to mixer which reads values from a YAML file
and makes them available to the rest of the server. This can be used to
quickly change server behavior on a per-environment basis.

The feature flag file location is passed in via the
`--feature_flags_path` command line flag. When no file is given, default
flag values are used (so Custom DC will fall back to default flag
values).

A script and wrapper Cloud Build action (defined in this PR but not yet
enabled in GCP) will update the ConfigMap from the latest files in
`deploy/featureflags`. A tool called
[Reloader](https://github.com/stakater/Reloader) (configured by the
`install_reloader` script in this PR in conjunction with an annotation
added to the Helm chart) watches for feature flag ConfigMap changes and
restarts mixer pods when changes are detected. Restarts are necessary
for ConfigMap changes to take effect.

**RISK:** If the binary and the ConfigMap become incompatible (a flag
type changes or a flag value fails custom validation rules), jobs will
fail to start.
- Mitigation: A pre-merge hook ensures that flag parsing always succeeds
before changes are merged, and the same test runs before changes are
deployed with Cloud Build (action defined in this PR but not enabled
yet).
- However, running `deploy_flags.sh` locally or deploying a server
binary update manually may still result in crashes.
- Note that mixer jobs are able to start if a flag is missing or an
extraneous flag is found. Only type mismatches or custom validation
failures will cause crashes.

To demonstrate custom validation rules and to set up for a switchover
from command line flags to feature flags for some existing flags,
`EnableV3` and `V3MirrorFraction` have been added to the flag configs in
this PR.

Tested on mixer dev. Not yet tested on a deployment of mixer bundled
with website, and additional config files for those environments will
have to be added.