Initial usage logger (#1608)
Basic usage logger that writes logs to `stdout` that will be ingested by
Cloud Logging. More fields can be added later -- this is the
foundational logger that makes a new `usageLogger` package in mixer and
builds the support for date formatting and result processing.

Current version includes the following:
- Information about every facet used in the result for each `statVar`
queried
- Query type, e.g. if the query is looking for the actual value of a
stat var or if it's looking for facet information or an existence check
- A metadata tag for which feature the query originates from
- This is especially valuable to have pushed out because having the
metadata header printed to logs will help me test to confirm that the
headers I'm passing in from other features are reaching Mixer correctly
- I also pass a "x-remote" flag into `FetchRemote` to track if a call is
coming to mixer remotely via a Custom DC. This is separate from the
`x-surface` flag because calls might come from another surface via
Custom DC -- for example, if a call is made to the One MCP, I want to
record that the call came from a custom DC but also from the MCP server.
In my logs, this will be represented by an attribute like `{ remote:
boolean, surface: string }`

The logs are written behind the `WriteUsageLogs` feature flag that is
set to 100% in dev, 10% in autopush, and 0% in staging and prod.

Fields to add in the future:
- [ ] Place type