[Feature Flags] Prevent prod flag changes outside allowed hours (#1630)
Adds a GitHub workflow that fails when:
- A file matching the pattern `deploy/featureflags/prod.*` has changed
- The current time is not Mon - Thurs, 8 AM - 5 PM in California

So this check will run immediately before merge, it will need to work in
conjunction with changes to master branch protections in the mixer repo:
- Branch `master` will require merge queue
- To reduce cognitive overhead for the team, the merge queue will only
allow one PR at a time.
- To avoid running the full test suite again, a no-op Cloud Build
trigger definition is added. This will be used by a new trigger in
datcom-ci that is *also* called `mixer-pull-request`, but triggers on
push to a merge queue branch targeting `master` instead (matches the
pattern `^gh-readonly-queue/master/.*`). This also works around a
limitation of Google Cloud Build triggers where they can only trigger
*either* on pull request *or* on push to branch. Without the new
trigger, merge queue will wait for the required status check
`mixer-pull-request` and eventually time out because it is never
triggered. If running the full test suite again as part of merge queue
checks is desirable, the second `mixer-pull-request` trigger can be
modified to also use `cloudbuild.test.yaml`.