[V3 Mirroring] Ignore expected field diffs (#1655)
Most significantly, the facet ID generation mechanism has changed, so
keys need to be normalized or ignored for comparison.

Same with the next token generation mechanism.

This is a roll-forward of #1646, which was rolled back in #1650. The
rollback was an attempt to fix autopush builds, and the fix ended up
being unrelated to this PR.