Make node page size modifiable (#1657)
This PR adds a field for custom page size. (The PR doesn't yet change
the page size, which will be a follow up)

This is so internal requests can set their own page size, for example
the in-memory cache at startup.

For backward compatibility, I didn't modify the existing NodeRequest
proto, but eventually the page size can be part of the request
(https://google.aip.dev/158). If/when we move to v3 fully, we can add
this there instead.

(Note that at the moment only the spanner datasource supports custom
pagination)