Create interface for SpannerClient (#1671)
This PR creates an interface for the SpannerClient. The previous
SpannerClient is renamed spannerDatabaseClient and implements the
interface.

This is part of a refactor with the goal of creating a separate
spanner+redis implementation which will also implement the SpannerClient
interface and which will include caching for stale reads. This is mostly
a stylistic decision to help isolate the spanner database interactions
from the caching interactions.

This PR also cleans up some constants