Convert the `WriteUsageLogs` feature flag to a command line flag for permanent use (#1678)
Removes the `WriteUsageLogs` flag from the V2 handler and the feature
flag package now that it is rolled out to 100%. This will be followed by
a second PR removing it from the feature flag configs.

Because we don't want to write usage logs automatically for custom DC
deployments, this feature flag is replaced by a command line flag that
is set to true in the deployment config here. This flag is NOT added to
custom DC scripts so they aren't written.

Although the logs can in the future be published as an option for custom
DCs, they aren't yet ready to be used publicly and might incur
unintended Cloud Logging costs if added accidentally.

This also updates the test script in `test/setup.go` to set the flag to
false for all tests, and sets it to true in the
`TestV2Observation_UsageLog` test that explicitly tests the logger.