Add flag-gated stale reads to spanner (no cache) (#1679)
This PR adds stale reads for spanner, gated behind the flag
UseStaleReads.

The timestamp is fetched in the background from Spanner and stored
locally. Currently it's set to refetch every minute but we can adjust as
needed.

*Note about consistency* There is a small chance for inconsistent
results if the polling differs. I discussed this with Keyur offline and
we decided that given our data freshness expectations this is acceptable
and should happen very infrequently and shouldn't persist for long.

Note: some goldens were updated with fresh data, this is unrelated to
this change