Fix merging in BulkVariableGroupInfo (#1688)
This is to fix a bug where the absoluteName was not being returned for
base SVGs in a custom DC mixer (e.g. UN staging)

By default, the localResp should have a response for each node requested
(https://github.com/datacommonsorg/mixer/blob/3cbf49b09d2df9132318524c4a8e470a78872f1f/internal/server/v1/info/variable_group.go#L49),
so line 297 is generally true. So, this PR ensures that the absoluteName
and parentStatVarGroups fields are set for this case and overrides with
the remoteResp if missing.

Tested locally that the values are now getting set.